agregando las trduciones de en y es
