+++ src/index.scss
@@ -18,7 +18,6 @@
 * {
   margin: 0;
   padding: 0;
-  outline: none;
-  box-sizing: borhser-box;
+  box-sizing: border-box;
   line-height: normal;
 }
