+++ src/index.tsx
@@ -5,23 +5,14 @@ import './index.scss';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
+import resources from './utils/translations';
 import { I18nextProvider } from 'react-i18next';
 import i18next from 'i18next';
-
-import global_en from './translations/en/global.json';
-import global_es from './translations/es/global.json';
 
 i18next.init({
   interpolation: { escapeValue: false },
   lng: 'en',
-  resources: {
-    es: {
-      global: global_es,
-    },
-    en: {
-      global: global_en,
-    },
-  },
+  resources,
 });
 
 ReactDOM.render(
