+++ src/translations/en/Home.json
@@ -1,0 +1,6 @@
+{
+  "title": "Hi",
+  "sub-title": "I’m Jesus Chacon",
+  "description": "I specialize in front-end development. I am creative, proactive and committed with a passion for making things beautiful and efficient.",
+  "btn": "Hire me"
+}
