+++ src/translations/en/Portfolio.json
@@ -1,0 +1,18 @@
+{
+  "project-1":{
+    "description": "Shop for the selection of various samples of perfumes for men, women, unisex"
+  },
+  "project-2":{
+    "description": "A basic badge crud for users"
+  },
+  "project-3":{
+    "description": "Professional and hand-optimized landing hand construction using React and styled-components"
+  },
+  "project-4":{
+    "description": "Website using Vuejs to list the top 20 most popular cryptocurrencies on the market using the coincap api"
+  },
+  "project-5":{
+    "description": "A user search engine using the github api"
+  },
+  "btn-visit": "Visit site"
+}
