+++ src/translations/en/global.json
@@ -6,7 +6,7 @@
     "btn": "Hire me"
   },
   "aboutme":{
-    "description": "A self-driven software engineer. Able to effectively self-manage during independent projects, as well as collaborate in a team setting. I always give the best of myself in whatever I do.",
+    "description": "An autonomous frontend developoer. Able to effectively self-manage during independent projects, as well as collaborate in a team environment. I always do my best in everything I do.",
     "services": {
       "service-1":{
         "title": "Dedicated",
