+++ src/translations/es/Home.json
@@ -1,0 +1,6 @@
+{
+  "title": "Hola,",
+  "sub-title": "Soy jesus chacon",
+  "description": "Me especializo en desarrollo front-end. Soy creativo, proactivo y comprometido con una pasión por hacer las cosas hermosas y eficientes.",
+  "btn": "Contratame"
+}
