+++ src/translations/es/Portfolio.json
@@ -1,0 +1,18 @@
+{
+  "project-1":{
+    "description": "Tienda para la seleccion de diversas muestras de perfumes para hombre, mujer, unisex"
+  },
+  "project-2":{
+    "description": "Un crud basico de insignias para los usuarios"
+  },
+  "project-3":{
+    "description": "Construcción de mano de aterrizaje profesional y optimizada a mano usando React y styled-components"
+  },
+  "project-4":{
+    "description": "Sitio web usando Vuejs para listar las 20 mejores criptomonedas mas populares en el mercado usando la api de coincap"
+  },
+  "project-5":{
+    "description": "Un buscador de usuarios usando la api de github"
+  },
+  "btn-visit": "Visitar sitio"
+}
