+++ src/translations/es/Posts.json
@@ -1,0 +1,5 @@
+{
+  "post-1": {
+    "description": ""
+  }
+}
