+++ src/translations/es/global.json
@@ -1,12 +1,12 @@
 {
   "header":{
-    "title": "Hi",
+    "title": "Hi,",
     "sub-title": "I’m Jesus Chacon",
     "description": "I specialize in front-end development. I am creative, proactive and committed with a passion for making things beautiful and efficient.",
     "btn": "Hire me"
   },
   "aboutme":{
-    "description": "A self-driven software engineer. Able to effectively self-manage during independent projects, as well as collaborate in a team setting. I always give the best of myself in whatever I do.",
+    "description": "An autonomous frontend developoer. Able to effectively self-manage during independent projects, as well as collaborate in a team environment. I always do my best in everything I do.",
     "services": {
       "service-1":{
         "title": "Dedicated",
@@ -28,18 +28,18 @@
       "description": "I am currently studying the 7th semester of the computer science engineering degree."
     },
     "platzi":{
-      "title": "",
+      "title": "Platzi",
       "description": "I had the opportunity to be benefited in the 2nd round of the thousand scholarships that were awarded to Venezuelans by platzi in 2019 where I have so far 30 javascript and react courses"
     },
     "udemy":{
-      "title": "",
+      "title": "Udemy",
       "description": "I decided to invest in myself, so I have acquired some react and vuejs courses on udemy to be able to expand my knowledge in other javascript frameworks with the aim of becoming a fullstack."
     },
     "btn": "Learn More"
   },
   "experience":{
     "Villartechnologies":{
-      "description": "Be part of the agency that is dedicated to the development of different web applications, be part of the frontend team using some technologies like Angular 8, HTML, CSS JS, Bootstrap and Microservices like Docker."
+      "description": "I was part of the agency that is dedicated to the development of different web applications, I was in the frontend team using some technologies such as Angular 8, HTML, CSS JS, Bootstrap and Microservices such as Docker."
     },
     "Freelancer-in-Upwork":{
       "description": "I worked as a freelance frontend developer using the upwork platform where I had the opportunity to use technologies like ReactJs, Redux, Redux-thunk and template development using HTML, CSS, SCSS, JS."
