+++ src/utils/translations.tsx
@@ -1,0 +1,42 @@
+/**
+ * Spanish translations
+ */
+
+import Home_ES from '../translations/es/Home.json';
+import About_ES from '../translations/es/About.json';
+import Education_ES from '../translations/es/Education.json';
+import Experience_ES from '../translations/es/Experience.json';
+import Portfolio_ES from '../translations/es/Portfolio.json';
+import Posts_ES from '../translations/es/Posts.json';
+
+/**
+ * English translations
+ */
+
+import Home_EN from '../translations/en/Home.json';
+import About_EN from '../translations/en/About.json';
+import Education_EN from '../translations/en/Education.json';
+import Experience_EN from '../translations/en/Experience.json';
+import Portfolio_EN from '../translations/en/Portfolio.json';
+import Posts_EN from '../translations/en/Posts.json';
+
+const resources = {
+  es: {
+    Home: Home_ES,
+    About: About_ES,
+    Education: Education_ES,
+    Experience: Experience_ES,
+    Portfolio: Portfolio_ES,
+    Posts: Posts_ES,
+  },
+  en: {
+    Home: Home_EN,
+    About: About_EN,
+    Education: Education_EN,
+    Experience: Experience_EN,
+    Portfolio: Portfolio_EN,
+    Posts: Posts_EN,
+  },
+};
+
+export default resources;
